OSIS-4855 Move file to correct app + remove unused files
