--- base/forms/education_group/common.py
@@ -1,49 +1,0 @@
-##############################################################################
-#
-#    OSIS stands for Open Student Information System. It's an application
-#    designed to manage the core business of higher education institutions,
-#    such as universities, faculties, institutes and professional schools.
-#    The core business involves the administration of students, teachers,
-#    courses, programs and so on.
-#
-#    Copyright (C) 2015-2019 Université catholique de Louvain (http://www.uclouvain.be)
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
-#    the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    A copy of this license - GNU General Public License - is available
-#    at the root of the source code of this program.  If not,
-#    see http://www.gnu.org/licenses/.
-#
-##############################################################################
-from django import forms
-from django.conf import settings
-from django.utils import translation
-from django.utils.translation import gettext_lazy as _
-
-from base.models import campus
-
-
-class MainCampusChoiceField(forms.ModelChoiceField):
-    def __init__(self, queryset, *args, **kwargs):
-        queryset = campus.find_main_campuses()
-        super().__init__(queryset, *args, **kwargs)
-
-
-# TODO: Only used in program_management/ ==> Move to it ?
-class SelectLanguage(forms.Form):
-    language = forms.ChoiceField(widget=forms.RadioSelect,
-                                 choices=settings.LANGUAGES,
-                                 label=_('Select a language'),
-                                 required=False)
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.initial['language'] = translation.get_language()
