--- base/forms/education_group/search/quick_search.py
@@ -1,91 +1,0 @@
-# ############################################################################
-#  OSIS stands for Open Student Information System. It's an application
-#  designed to manage the core business of higher education institutions,
-#  such as universities, faculties, institutes and professional schools.
-#  The core business involves the administration of students, teachers,
-#  courses, programs and so on.
-#
-#  Copyright (C) 2015-2019 Université catholique de Louvain (http://www.uclouvain.be)
-#
-#  This program is free software: you can redistribute it and/or modify
-#  it under the terms of the GNU General Public License as published by
-#  the Free Software Foundation, either version 3 of the License, or
-#  (at your option) any later version.
-#
-#  This program is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
-#
-#  A copy of this license - GNU General Public License - is available
-#  at the root of the source code of this program.  If not,
-#  see http://www.gnu.org/licenses/.
-# ############################################################################
-from django import forms
-from django.utils.translation import gettext_lazy as _, pgettext_lazy
-from django_filters import FilterSet, filters, OrderingFilter
-
-from base.models.academic_year import AcademicYear
-from education_group.models.group_year import GroupYear
-
-
-class QuickGroupYearFilter(FilterSet):
-    academic_year = filters.ModelChoiceFilter(
-        queryset=AcademicYear.objects.all(),
-        to_field_name="year",
-        required=False,
-        label=_('Ac yr.'),
-        empty_label=pgettext_lazy("plural", "All"),
-    )
-    acronym = filters.CharFilter(
-        field_name="acronym",
-        lookup_expr='icontains',
-        max_length=40,
-        required=False,
-        label=_('Acronym/Short title'),
-    )
-    partial_acronym = filters.CharFilter(
-        field_name="partial_acronym",
-        lookup_expr='icontains',
-        max_length=40,
-        required=False,
-        label=_('Code'),
-    )
-    title = filters.CharFilter(
-        field_name="title_fr",
-        lookup_expr='icontains',
-        max_length=255,
-        required=False,
-        label=_('Title')
-    )
-
-    ordering = OrderingFilter(
-        fields=(
-            ('academic_year__year', 'academic_year'),
-            ('acronym', 'acronym'),
-            ('partial_acronym', 'code'),
-            ('title', 'title'),
-        ),
-        widget=forms.HiddenInput
-    )
-
-    class Meta:
-        model = GroupYear
-        fields = [
-            'acronym',
-            'title',
-            'academic_year',
-            'partial_acronym'
-        ]
-
-    def __init__(self, *args, initial=None, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.queryset = self.get_queryset()
-        if initial:
-            self.form.fields["academic_year"].initial = initial["academic_year"]
-
-    def get_queryset(self):
-        # Need this close so as to return empty query by default when form is unbound
-        if not self.data:
-            return GroupYear.objects.none()
-        return GroupYear.objects.all()
